extract Composables to files and simplify passed parameters
