+++ app/src/main/java/com/nextcloud/talk/contacts/components/SetStatusBarColor.kt
@@ -1,0 +1,37 @@
+/*
+ * Nextcloud Talk - Android Client
+ *
+ * SPDX-FileCopyrightText: 2024 Sowjanya Kota <sowjanya.kch@gmail.com>
+ * SPDX-FileCopyrightText: 2025 Marcel Hibbe <dev@mhibbe.de>
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package com.nextcloud.talk.contacts.components
+
+import android.app.Activity
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.ui.graphics.toArgb
+
+import androidx.compose.ui.platform.LocalView
+import androidx.compose.ui.res.colorResource
+import androidx.core.view.WindowCompat
+import com.nextcloud.talk.R
+
+@Composable
+fun SetStatusBarColor() {
+    val view = LocalView.current
+    val isDarkMod = isSystemInDarkTheme()
+    val statusBarColor = colorResource(R.color.bg_default).toArgb()
+
+    DisposableEffect(isDarkMod) {
+        val activity = view.context as Activity
+        activity.window.statusBarColor = statusBarColor
+
+        WindowCompat.getInsetsController(activity.window, activity.window.decorView).apply {
+            isAppearanceLightStatusBars = !isDarkMod
+        }
+        onDispose { }
+    }
+}
